+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # verbecc Changelog
 
-- 1.11.6 [TBR]
-  - Fixed Voseo conjugation for irregular verb `ser`
+- 1.11.6 [26 October 2025]
+  - Fixed Voseo conjugation for irregular verb `ser` for the subjuntivo (no vowel accents)
   - Misc. cleanup
   - Promoted applicable `Inflector` methods to public